vaulet/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
Pillow 
django-tailwind
django-filter
sqlalchemy
phidata
google-generativeai
duckduckgo-search
celery==5.3.6
django-celery-beat==2.5.0
redis